Refactor DbSpectral in prep for sharing with DbImgSampler
