+++ DbImgTools/DbSpectral/main.cpp
