+++ DbImgTools/README.md
@@ -1,0 +1,39 @@
+# DbImgTools Chugins
+
+## Intro
+
+DbImgTools is a collection of [ChucK](https://chuck.stanford.edu/) chugins 
+focused on the use of image files to control the synthesis parameters.
+
+[DbSpectral](DbSpectral/README.md) is a time-varying audio filter described
+by an image file.
+
+[DbImgSampler](DbImgSampler/README.md) base be used to sample an image
+for the purposes of producing control parameters for other audio processing
+tools.
+
+## License
+
+DbImgTools are open-source.  Sub-components are subject to their own
+licenses.  Unless otherwise stated, this code is subject to the
+to the MIT license:
+
+Copyright 2022 Dana Batali
+
+Permission is hereby granted, free of charge, to any person obtaining a copy 
+of this software and associated documentation files (the "Software"), to 
+deal in the Software without restriction, including without limitation the 
+rights to use, copy, modify, merge, publish, distribute, sublicense, and/or 
+sell copies of the Software, and to permit persons to whom the Software is 
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in 
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE 
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER 
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, 
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE 
+SOFTWARE.
