+++ DbImgTools/lib/fftsg.cpp
