+++ DbImgTools/lib/fftsg.h
