+++ DbImgTools/lib/image.cpp
