+++ DbImgTools/lib/image.h
