+++ DbImgTools/lib/ringbuffer.h
