+++ DbImgTools/lib/stb_image_resize.h
