+++ DbImgTools/test/imgerr.ck
